Split code and package it
