+++ pyproject.toml
@@ -1,0 +1,23 @@
+[build-system]
+requires = ["setuptools>=61.0"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "selectk"
+version = "0.0.1"
+authors = [
+  { name="Marcelo Henriques Cortez", email="marcelohcortez@gmail.com" },
+  { name="Carlos Pinto Machado", email="cpmachado@protonmail.com" },
+]
+description = "methods to select kth element of a python list"
+readme = "README.md"
+requires-python = ">=3.10"
+classifiers = [
+    "Programming Language :: Python :: 3",
+    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
+    "Operating System :: OS Independent",
+]
+
+[project.urls]
+"Homepage" = "https://github.com/marcelohcortez/python_kth_largest"
+"Bug Tracker" = "https://github.com/marcelohcortez/python_kth_largest/issues"
