+++ tests/test_quickselect.py
@@ -1,4 +1,4 @@
-from quickselect import find_kth_largest_ele
+from selectk.quickselect import find_kth_largest_ele
 
 def test_find_kth_largest_ele():
     for i in range(10):
